removed tm convenience function
