+++ src/probnum/diffeq/odefiltsmooth/initialize/__init__.py
@@ -1,13 +1,13 @@
 """Initialisation procedures for ODE filters."""
 
 from ._initialize import InitializationRoutine
-from ._initialize_with_taylormode import initialize_odefilter_with_taylormode
 from ._runge_kutta import RungeKuttaInitialization
+from ._taylor_mode import TaylorModeInitialization
 
 __all__ = [
     "InitializationRoutine",
     "RungeKuttaInitialization",
-    "initialize_odefilter_with_taylormode",
+    "TaylorModeInitialization",
 ]
 
 
@@ -15,3 +15,4 @@ __all__ = [
 # Corrects links and module paths in documentation.
 InitializationRoutine.__module__ = "probnum.diffeq.odefiltsmooth.initialize"
 RungeKuttaInitialization.__module__ = "probnum.diffeq.odefiltsmooth.initialize"
+TaylorModeInitialization.__module__ = "probnum.diffeq.odefiltsmooth.initialize"
